Moved CardBenefitsSection to the Home folder
