+++ frontend/front/src/pages/Public/Pages/Home/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Container } from "@mui/material";
 import Partners from "./Partners";
-import CardBenefitsSection from "../../Components/CardBenefitsSection";
+import CardBenefitsSection from "./CardBenefitsSection";
 import CardLinksSection from "./CardLinksSection";
 
 import surveyorImage from "../../../../assets/images/surveyor.jpg";
